Make AbstractOffHeapArray Java 9+ compatible.
Java 9 moved sun.misc.Cleaner to jdk.internal.ref.Cleaner. In order
for this library to be compatible with both Java 8 and Java 9+
runtimes, we need to decide between those at runtime, which
unfortunately means using reflection.

It is worth noting that in Java 9 there was also a public API
introduced for using a cleaner (at java.lang.ref.Cleaner), but that
API is a bit more involved and so it isn't really worth using via
reflection as of right now. If this library ever upgrades to Java 9+
compatibility only, then all this reflection stuff should be replaced
with that instead of the jdk.internal.ref.Cleaner.